Fix all the tests
