+++ src/Core/tests/DeviceTests/Handlers/Label/LabelHandlerTests.Android.cs
@@ -1,9 +1,7 @@
 using System.Threading.Tasks;
 using Android.Text;
 using Android.Widget;
-using Microsoft.Maui.DeviceTests.Stubs;
 using Microsoft.Maui.Handlers;
-using Xunit;
 
 namespace Microsoft.Maui.DeviceTests
 {
@@ -31,30 +29,5 @@ namespace Microsoft.Maui.DeviceTests
 
 		TextUtils.TruncateAt GetNativeLineBreakMode(LabelHandler labelHandler) =>
 			GetNativeLabel(labelHandler).Ellipsize;
-
-		[Fact(DisplayName = "LineBreakMode Initializes Correctly")]
-		public async Task LineBreakModeInitializesCorrectly()
-		{
-			var xplatLineBreakMode = LineBreakMode.TailTruncation;
-
-			var labelStub = new LabelStub()
-			{
-				LineBreakMode = xplatLineBreakMode
-			};
-
-			var expectedValue = TextUtils.TruncateAt.End;
-
-			var values = await GetValueAsync(labelStub, (handler) =>
-			{
-				return new
-				{
-					ViewValue = labelStub.LineBreakMode,
-					NativeViewValue = GetNativeLineBreakMode(handler)
-				};
-			});
-
-			Assert.Equal(xplatLineBreakMode, values.ViewValue);
-			Assert.Equal(expectedValue, values.NativeViewValue);
-		}
 	}
 }
